backend/requirements.txt
asgiref==3.8.1
crispy-bootstrap4==2024.1
Django==5.1.1
django-crispy-forms==2.3
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
fontawesomefree==6.6.0
psycopg==3.1.12
psycopg-binary==3.1.12
PyJWT==2.9.0
python-decouple==3.8
sqlparse==0.5.1
typing_extensions==4.12.2
psycopg[binary]==3.1.12
dj-rest-auth==6.0.0
channels[daphne]
channels_redis
pyotp
django-cors-headers==3.13.0
django-truncate
python-dotenv
